// frontend/src/api/generated/models/interface-file-info.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface InterfaceFileInfo
 */
export interface InterfaceFileInfo {
    /**
     * CSVファイル名
     * @type {string}
     * @memberof InterfaceFileInfo
     */
    'fileName': string;
    /**
     * ファイルサイズ（バイト）
     * @type {number}
     * @memberof InterfaceFileInfo
     */
    'fileSize': number;
    /**
     * Blob Storageへの配置日時
     * @type {string}
     * @memberof InterfaceFileInfo
     */
    'createdAt': string;
}