// frontend/src/api/generated/models/interface-file-list-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { InterfaceFileInfo } from './interface-file-info';

/**
 * 
 * @export
 * @interface InterfaceFileListResponse
 */
export interface InterfaceFileListResponse {
    /**
     * 
     * @type {Array<InterfaceFileInfo>}
     * @memberof InterfaceFileListResponse
     */
    'files': Array<InterfaceFileInfo>;
    /**
     * ファイル件数
     * @type {number}
     * @memberof InterfaceFileListResponse
     */
    'totalCount': number;
}