// frontend/src/api/generated/models/interface-validation-result.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { InterfaceFileError } from './interface-file-error';
// May contain unused imports in some cases
// @ts-ignore
import type { InterfaceValidationRow } from './interface-validation-row';

/**
 * 
 * @export
 * @interface InterfaceValidationResult
 */
export interface InterfaceValidationResult {
    /**
     * CSVファイル名
     * @type {string}
     * @memberof InterfaceValidationResult
     */
    'fileName': string;
    /**
     * 総データ行数（ヘッダ除く）
     * @type {number}
     * @memberof InterfaceValidationResult
     */
    'totalRows': number;
    /**
     * バリデーション成功行数
     * @type {number}
     * @memberof InterfaceValidationResult
     */
    'successCount': number;
    /**
     * バリデーションエラー行数
     * @type {number}
     * @memberof InterfaceValidationResult
     */
    'errorCount': number;
    /**
     * 
     * @type {InterfaceFileError}
     * @memberof InterfaceValidationResult
     */
    'fileError'?: InterfaceFileError;
    /**
     * エラー行のみ含む
     * @type {Array<InterfaceValidationRow>}
     * @memberof InterfaceValidationResult
     */
    'rows': Array<InterfaceValidationRow>;
}