// frontend/src/api/generated/models/interface-import-result.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ImportMode } from './import-mode';

/**
 * 
 * @export
 * @interface InterfaceImportResult
 */
export interface InterfaceImportResult {
    /**
     * 取り込み成功行数
     * @type {number}
     * @memberof InterfaceImportResult
     */
    'successCount': number;
    /**
     * バリデーションエラー行数
     * @type {number}
     * @memberof InterfaceImportResult
     */
    'errorCount': number;
    /**
     * 
     * @type {ImportMode}
     * @memberof InterfaceImportResult
     */
    'mode': ImportMode;
    /**
     * 取り込みステータス
     * @type {string}
     * @memberof InterfaceImportResult
     */
    'status': InterfaceImportResultStatusEnum;
}

export const InterfaceImportResultStatusEnum = {
    Completed: 'COMPLETED',
    Discarded: 'DISCARDED',
    Failed: 'FAILED'
} as const;

export type InterfaceImportResultStatusEnum = typeof InterfaceImportResultStatusEnum[keyof typeof InterfaceImportResultStatusEnum];